Only run scheduled tests on GitHub for datalad/datalad
This adds a repository conditional to the jobs in the "Test" GitHub workflow,
which is triggered on schedule. As such, it should be disabled by default in
forks, but e.g. manually pushing the main branch to a fork created before the
action was added will cause it to become enabled.

Based on:
https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#example-only-run-job-for-specific-repository

See also:
https://docs.github.com/en/actions/managing-workflow-runs-and-deployments/managing-workflow-runs/disabling-and-enabling-a-workflow