test: xfail test_subsuperdataset_save on newer gits
The whole point for that "permissions" trick was to ensure that
git does not go into submodules and thus we would be "optimized" and
not waste cycles where not desired.  But the very fact that it errors
out points that it is not the case.

So this provides only taming down for
  https://github.com/datalad/datalad/issues/7681
and for a proper solution we better first wait on what git gurus say in
  https://lore.kernel.org/git/ZzyyQpJw-kxJA7IS@bilena/T/#u