fix: submodule property `datalad-url` used wrong submodule name
Instead of the relative path in POSIX conventions, it used the
relative path in platform conventions. On windows this led
to a dangling submodule property, detached from the actual
submodule record, whenever a submodule was added in a 2nd-level
subdirectory. This led to warnings from routines that parsed
the submodule.

This change force-converts the name/path to POSIX conventions.